+++ refinedstorage-emi-integration-common/src/main/java/com/refinedmods/refinedstorage/emi/common/RefinedStorageEmiPlugin.java
@@ -13,7 +13,9 @@ public class RefinedStorageEmiPlugin implements EmiPlugin {
         registry.addGenericExclusionArea(new EmiExclusionAreaImpl());
         registry.addGenericStackProvider(new GridEmiStackProvider());
         registry.addGenericStackProvider(new ResourceEmiStackProvider());
-        registry.addGenericDragDropHandler(new EmiDragDropHandlerImpl());
-        registry.addRecipeHandler(Menus.INSTANCE.getCraftingGrid(), new EmiRecipeHandlerImpl());
+        registry.addGenericDragDropHandler(new ResourceEmiDragDropHandler());
+        registry.addGenericDragDropHandler(new FilterEmiDragDropHandler());
+        registry.addRecipeHandler(Menus.INSTANCE.getCraftingGrid(), new CraftingGridEmiRecipeHandler());
+        registry.addRecipeHandler(Menus.INSTANCE.getPatternGrid(), new PatternGridEmiRecipeHandler());
     }
 }