+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+-   Recipe transfer support for the Pattern Grid.
+
 ## [0.3.2] - 2024-07-26
 
 ### Fixed